[refactor] checkin into modules folder, update README todos to reflect what has changed.
